agentnotify: move agent notification into their own package
